dom/file/ipc/tests/test_ipcBlob_emptyMultiplex.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test an empty IPCBlob together with other parts</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script type="text/javascript">

function checkContent(msg, content) {
  return new Promise(resolve => {
    let fr = new FileReader();
    fr.readAsText(new Blob(content));
    fr.onloadend = () => {
      is(fr.result, "Hello world!", "The content matches: " + msg);
      resolve();
    };
  });
}

SimpleTest.waitForExplicitFinish();

let url = SimpleTest.getTestFileURL("script_file.js");

let script = SpecialPowers.loadChromeScript(url);
script.addMessageListener("emptyfile.opened", message => {
  checkContent("middle", ["Hello ", message.file, "world!"]).
  then(() => checkContent("begin", [message.file, "Hello world!"])).
  then(() => checkContent("end", ["Hello world!", message.file])).
  then(() => checkContent("random", [message.file, message.file, "Hello world!", message.file])).
  then(() => checkContent("random 2", [message.file, message.file, "Hello ",
                                       message.file, "world", message.file,
                                       message.file, "!", message.file, "",
                                       message.file, message.file])).
  then(SimpleTest.finish);
});

script.sendAsyncMessage("emptyfile.open");

</script>
</pre>
</body>
</html>
